Consolidate Page classes, rework app.current, app.previous
