+++ js/src/admin/components/DashboardPage.js
@@ -1,4 +1,4 @@
-import Page from './Page';
+import Page from '../../common/components/Page';
 import StatusWidget from './StatusWidget';
 
 export default class DashboardPage extends Page {
