+++ js/src/common/components/ModalManager.js
@@ -43,7 +43,7 @@ export default class ModalManager extends Component {
     this.showing = true;
     this.component = component;
 
-    if (app.current) app.current.retain = true;
+    // if (app.current) app.current.retain = true;
 
     m.redraw(true);
