+++ js/src/common/utils/subclassOf.js
@@ -2,5 +2,5 @@
  * Check if class A is a subclass of class B.
  */
 export default function subclassOf(A, B) {
-  return A == B || A.prototype instanceof B;
+  return A && (A == B || A.prototype instanceof B);
 }
