+++ js/src/forum/ForumApplication.js
@@ -16,6 +16,7 @@ import Navigation from '../common/components/Navigation';
 import NotificationListState from './states/NotificationListState';
 import GlobalSearchState from './states/GlobalSearchState';
 import DiscussionListState from './state/DiscussionListState';
+import subclassOf from '../common/utils/subclassOf';
 
 export default class ForumApplication extends Application {
   /**
@@ -160,7 +161,7 @@ export default class ForumApplication extends Application {
    * @return {Boolean}
    */
   viewingDiscussion(discussion) {
-    return this.current instanceof DiscussionPage && this.current.discussion === discussion;
+    return subclassOf(this.current, DiscussionPage) && this.currentData.discussion === discussion;
   }
 
   /**
