+++ js/src/forum/components/RenameDiscussionModal.js
@@ -57,7 +57,7 @@ export default class RenameDiscussionModal extends Modal {
         .save({ title })
         .then(() => {
           if (app.viewingDiscussion(this.discussion)) {
-            app.current.stream.update();
+            app.currentData.stream.update();
           }
           m.redraw();
           this.hide();
