+++ js/src/forum/components/UserPage.js
@@ -71,6 +71,8 @@ export default class UserPage extends Page {
   show(user) {
     this.user = user;
 
+    app.currentData = { user };
+
     app.setTitle(user.displayName());
 
     m.redraw();
